Parsing is now working really well, reformating made the logic of spaces really simpler en more robust
