+++ srcs/expand_reduce.c
@@ -6,7 +6,7 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/17 17:19:43 by smagdela          #+#    #+#             */
-/*   Updated: 2022/03/21 12:00:05 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/03/21 14:20:13 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -151,7 +151,7 @@ bool	reduce_words(t_token *elem, size_t end)
 
 	if (elem == NULL || elem->index > end)
 	{
-		printf("elem ptr = %p\nelem index = %lu\nend = %lu\n", elem, elem->index, end);
+//		printf("elem ptr = %p\nelem index = %lu\nend = %lu\n", elem, elem->index, end);
 		return (false);
 	}
 	cmd = (char **)malloc(sizeof(char *) * (end - elem->index + 2));
